add ability to add and filter heroes

+++ src/components/heroesList/HeroesList.js
@@ -7,15 +7,14 @@ import HeroesListItem from "../heroesListItem/HeroesListItem";
 import Spinner from '../spinner/Spinner';
 
 const HeroesList = () => {
-    const {heroes, heroesLoadingStatus} = useSelector(state => state);
+    const {filteredHeroes, heroesLoadingStatus} = useSelector(state => state);
     const dispatch = useDispatch();
     const {request, reqToDelete} = useHttp();
 
     useEffect(() => {
         dispatch(heroesFetching());
         request("http://localhost:3001/heroes")
-            .then(data => {dispatch(heroesFetched(data))
-            console.log(data)})
+            .then(data => {dispatch(heroesFetched(data))})
             .catch(() => dispatch(heroesFetchingError()))
 
         // eslint-disable-next-line
@@ -36,16 +35,21 @@ const HeroesList = () => {
     }
 
     const renderHeroesList = (arr) => {
-        if (arr.length === 0) {
-            return <h5 className="text-center mt-5">Героев пока нет</h5>
+        if (!Array.isArray(arr)) {
+          return <h5 className="text-center mt-5">Invalid hero list</h5>;
         }
+      
+        if (arr.length === 0) {
+          return <h5 className="text-center mt-5">No heroes found</h5>;
+        }
+      
+        return arr.map(({ id, ...props }) => {
+          return <HeroesListItem key={id} id={id} onDelete={deleteItem} {...props} />;
+        });
+      };
+      
 
-        return arr.map(({id, ...props}) => {
-            return <HeroesListItem key={id} id={id} onDelete={deleteItem} {...props}/>
-        })
-    }
-
-    const elements = renderHeroesList(heroes);
+    const elements = renderHeroesList(filteredHeroes);
     return (
         <ul>
             {elements}
